added tutorial, data, readme, and fixed imports

+++ tissue/main.py
@@ -11,7 +11,7 @@ import anndata as ad
 import warnings
 import os
 
-from .utils import nan_weighted_std
+from tissue.utils import nan_weighted_std
 
 
 def load_paired_datasets (spatial_counts, spatial_loc, RNAseq_counts, spatial_metadata = None,
@@ -61,7 +61,7 @@ def load_spatial_data (spatial_counts, spatial_loc, spatial_metadata=None,
     df = df.loc[genes_prevalence > min_gene_prevalence_spatial, :]
     del genes_prevalence
     # create AnnData
-    spatial_adata = ad.AnnData(X=df)
+    spatial_adata = ad.AnnData(X=df, dtype='float64')
     spatial_adata.obs_names = df.index.values
     spatial_adata.obs_names = spatial_adata.obs_names.astype(str)
     spatial_adata.var_names = df.columns
@@ -97,7 +97,7 @@ def load_rnaseq_data (RNAseq_counts, min_cell_prevalence_RNAseq = 0.0, min_gene_
     df = df.loc[genes_prevalence > min_gene_prevalence_RNAseq, :]
     del genes_prevalence
     # create AnnData
-    RNAseq_adata = ad.AnnData(X=df.T)
+    RNAseq_adata = ad.AnnData(X=df.T, dtype='float64')
     RNAseq_adata.obs_names = df.T.index.values
     RNAseq_adata.var_names = df.T.columns
     del df
@@ -450,7 +450,7 @@ def spage_impute (spatial_adata, RNAseq_adata, genes_to_predict, **kwargs):
     Runs SpaGE gene imputation
     '''
     #sys.path.append("Extenrnal/SpaGE-master/")
-    from SpaGE.main import SpaGE
+    from tissue.SpaGE.main import SpaGE
     
     # transform adata in spage input data format
     if isinstance(spatial_adata.X,np.ndarray):
